# docs/swagger.yaml
basePath: /api
definitions:
...
  apierror.ApiError:
    properties:
      code:
        type: integer
      error: {}
    type: object
  db.Assignment:
    properties:
      course_id:
        type: integer
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      description:
        $ref: '#/definitions/pgtype.Text'
      due_date:
        $ref: '#/definitions/pgtype.Timestamp'
      id:
        type: integer
      max_attempts_per_task:
        $ref: '#/definitions/pgtype.Int4'
      start_date:
        $ref: '#/definitions/pgtype.Timestamp'
      title:
        type: string
      updated_at:
        $ref: '#/definitions/pgtype.Timestamp'
    type: object
  db.Course:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      description:
        $ref: '#/definitions/pgtype.Text'
      id:
        type: integer
      join_code:
        type: string
      name:
        type: string
      teacher_id:
        type: integer
      updated_at:
        $ref: '#/definitions/pgtype.Timestamp'
    type: object
...
  models.AddTaskToAssignmentDTO:
    properties:
      order_index:
        minimum: 0
        type: integer
      task_id:
        type: integer
      weight:
        maximum: 5
        minimum: 0.1
        type: number
    required:
    - task_id
    type: object
  models.AssignmentResponse:
    properties:
      course:
        $ref: '#/definitions/db.Course'
      course_id:
        type: integer
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      description:
        $ref: '#/definitions/pgtype.Text'
      due_date:
        $ref: '#/definitions/pgtype.Timestamp'
      id:
        type: integer
      max_attempts_per_task:
        $ref: '#/definitions/pgtype.Int4'
      start_date:
        $ref: '#/definitions/pgtype.Timestamp'
      tasks:
        items:
          $ref: '#/definitions/models.TaskWithOrder'
        type: array
      title:
        type: string
      total_tasks:
        type: integer
      updated_at:
        $ref: '#/definitions/pgtype.Timestamp'
    type: object
  models.AssignmentStats:
    properties:
      average_grade:
        type: number
      completed_count:
        type: integer
      submissions_count:
        type: integer
      total_students:
        type: integer
    type: object
...
  models.CreateAssignmentDTO:
    properties:
      course_id:
        type: integer
      description:
        maxLength: 1000
        type: string
      due_date:
        type: string
      max_attempts_per_task:
        maximum: 100
        minimum: 1
        type: integer
      start_date:
        type: string
      title:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - course_id
    - title
    type: object
...
  models.TaskWithOrder:
    properties:
      course_id:
        $ref: '#/definitions/pgtype.Int4'
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      description:
        type: string
      difficulty:
        type: integer
      id:
        type: integer
      is_public:
        type: boolean
      order_index:
        type: integer
      passes:
        type: integer
      score:
        type: number
      title:
        type: string
      topics:
        items:
          type: integer
        type: array
      updated_at:
        $ref: '#/definitions/pgtype.Timestamp'
      user_id:
        type: integer
      weight:
        type: number
    type: object
  models.UpdateAssignmentDTO:
    properties:
      description:
        maxLength: 1000
        type: string
      due_date:
        type: string
      max_attempts_per_task:
        maximum: 100
        minimum: 1
        type: integer
      start_date:
        type: string
      title:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - title
    type: object
...
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Int4:
    properties:
      int32:
        type: integer
      valid:
        type: boolean
    type: object
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.Timestamp:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        description: Time zone will be ignored when encoding to PostgreSQL.
        type: string
      valid:
        type: boolean
    type: object
...
host: localhost:4000
info:
  contact: {}
  title: Rankode
  version: "1.0"
paths:
  /assignments:
    get:
      consumes:
      - application/json
      description: Retrieves assignments for the current user (as teacher or student)
      parameters:
      - description: Filter by course ID
        in: query
        name: course_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of assignments
          schema:
            items:
              $ref: '#/definitions/db.Assignment'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: List assignments
      tags:
      - Assignments
    post:
      consumes:
      - application/json
      description: Creates a new assignment in a course (teacher only)
      parameters:
      - description: Assignment creation payload
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/models.CreateAssignmentDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created assignment
          schema:
            $ref: '#/definitions/db.Assignment'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "403":
          description: Forbidden (not teacher of course)
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Create a new assignment
      tags:
      - Assignments
  /assignments/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an assignment (teacher only)
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Assignment deleted successfully
        "403":
          description: Forbidden (not teacher of course)
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "404":
          description: Assignment not found
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Delete assignment
      tags:
      - Assignments
    get:
      consumes:
      - application/json
      description: Retrieves a single assignment by its ID with included tasks
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Assignment details
          schema:
            $ref: '#/definitions/models.AssignmentResponse'
        "404":
          description: Assignment not found
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Get assignment by ID with tasks
      tags:
      - Assignments
    put:
      consumes:
      - application/json
      description: Updates an existing assignment (teacher only)
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Assignment update payload
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAssignmentDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Assignment updated successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "403":
          description: Forbidden (not teacher of course)
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "404":
          description: Assignment not found
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Update assignment
      tags:
      - Assignments
  /assignments/{id}/stats:
    get:
      consumes:
      - application/json
      description: Retrieves statistics for an assignment (teacher only)
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Assignment statistics
          schema:
            $ref: '#/definitions/models.AssignmentStats'
        "403":
          description: Forbidden (not teacher of course)
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "404":
          description: Assignment not found
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Get assignment statistics
      tags:
      - Assignments
  /assignments/{id}/tasks:
    post:
      consumes:
      - application/json
      description: Adds an existing task to an assignment (teacher only)
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task addition payload
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.AddTaskToAssignmentDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Task added to assignment successfully
        "400":
          description: Bad request (task not in course, already added)
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "403":
          description: Forbidden (not teacher of course)
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "404":
          description: Assignment or task not found
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Add task to assignment
      tags:
      - Assignments
  /assignments/{id}/tasks/{taskId}:
    delete:
      consumes:
      - application/json
      description: Removes a task from an assignment (teacher only)
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Task removed from assignment successfully
        "403":
          description: Forbidden (not teacher of course)
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "404":
          description: Assignment or task not found
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Remove task from assignment
      tags:
      - Assignments
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"